+++ aegir/tools/system/mysql_hourly.sh
@@ -36,8 +36,8 @@ check_root() {
 }
 check_root
 
-rm -f /root/.pause_tasks_maint.cnf
-rm -f /root/.restrict_this_vm.cnf
+[ -e "/root/.pause_tasks_maint.cnf" ] && rm -f /root/.pause_tasks_maint.cnf
+[ -e "/root/.restrict_this_vm.cnf" ] && rm -f /root/.restrict_this_vm.cnf
 
 if [ -e "/root/.proxy.cnf" ]; then
   exit 0
